+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - fetchWithProgress takes an element, not ID, to work better with Svelte
+- Added `isActive()` method and new events `activate` and `no-new-route`
 
 ## 0.1.1